requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
datasets==4.2.0
dill==0.3.6
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.75.1
h11==0.16.0
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
Jinja2==3.1.6
Markdown==3.9
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.14
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==9.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
responses==0.18.0
safetensors==0.6.2
sentry-sdk==2.42.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.22.2
wcwidth==0.2.14
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
